requirements.txt
tabulate
pandas
mysql-connector-python
python-dotenv